# src/common/swagger.yaml
swagger: '2.0'
info:
  description: "LINS project REST api"
  version: "1.0"
  title: "LINS backend"
  contact:
    name: "Developer"
    url: "https://github.com/motorny"
    email: "motorny.nikita@gmail.com"
...
schemes:
  - "http"
paths:
...
  /comments:
    post:
      tags:
        - "comments"
      summary: "Add new comment"
      description: "Certain user adds comment to an item"
      operationId: "addComment"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "commentData"
          description: "Comment data"
          required: true
          schema:
            $ref: "#/definitions/Comment"
      responses:
        200:
          description: "Success"
        400:
          description: "Wrong item appeal"
      security:
        - JWT: []
  /comments/{id}:
    get:
      tags:
        - "comments"
      summary: "Get all comments for certain item"
      description: "getAllComments"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the item"
          reqired: true
          type: "integer"
          format: "integer"
      responses:
        200:
          description: "Success"
        412:
          description: "Item ID is not valid"
      security:
        - JWT: []

    delete:
      tags:
        - "comments"
      summary: "Delete comment"
      description: ""
      operationId: "deleteComment"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the comment"
          required: true
          type: "integer"
          format: "integer"
      responses:
        200:
          description: "Success"
        403:
          description: "Not authorized or token not valid"
        412:
          description: "Comment ID is not valid"
      security:
        - JWT: []

    put:
      tags:
        - "comments"
      summary: "Update existing comment"
      description: ""
      operationId: "editComment"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the comment"
          required: true
          type: "integer"
          format: "integer"
        - in: "body"
          name: "commentData"
          description: "Comment data"
          required: true
          schema:
            $ref: "#/definitions/Comment"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Comment"
        403:
          description: "Not authorized for this action or token not valid"
        412:
          description: "Comment ID is not valid"
      security:
        - JWT: []
securityDefinitions:
  JWT:
    type: apiKey
    in: header
    name: Authorization
definitions:
...
  Comment:
    type: "object"
    properties:
      title:
        title: "Title"
        type: "string"
        description: "Title for comment"
        example: "Best item in the world!"
        minLength: 1
      image:
        title: "Item's image"
        description: "base64 encoded image"
        type: "string"
      comment:
        title: "Commentary itself"
        type: "string"
        description: "Full text of the comment"
        example: "Used for three days. Very grateful for this item"
      user_id:
        title: "User id"
        type: "integer"
        description: "Id of the user, who left the comment"
      item_id:
        title: "Item id"
        type: "integer"
        description: "If of the item to which comment is written"
...